requirements.txt
-e .
aws_cdk.aws_s3
